Resolve quote too old confusion.
We try never to use `time` when dealing with contracts. This makes it too
difficult to reproduce the contract later.  Instead, we take
date_pricing as the proxy for "current time."

ALso, there is no need to check specifically for forward_starting.  This
is a test on the underlying.  It is someone else's job to figure out if
the date_pricing is correct for the contract time.