Note failure modes for buys and sells.
This does not cover the automated sales processes, but will let us
understand why customer-initiated transactions fail.

The timings are ommitted on failures, because it would alter the
baseline numbers. This may be worthy of reconsideration.

The `stats_stop` name might be slightly confusing now.  It does what is
says, but it also passes back the errors for the early returns, when
supplied.