Correct tick count for AggTicks.
On very sparse ticks we were accidentally leaving some ticks uncounted
in our insertion.  This problem is corrected with appropriate regression
tests to show the particular problem resolved (see the 45 and 60
structure tests)

The first insertion has to be slightly special-cased because we haven't
properly gotten into the flow of previous versus current tick.

The tick counts on the example prices have always varied as we change
the model.  Having reviewed the numbers manually (plus the regression
test), they seem much more likely to be correct than before.