Improve barrier moved messages.
Separate out the numbers a bit so it's clear what they are and where
they are used.

The numbers in the tags do not end up making sense.. and we might as
well include how far off they were for any other data gathering.

The localiazation makes some of this uglier and less DRY than one might
like, but it should get the job done.