Require exit tick for American expiries at the end.
We don't actually use this info, but if it is present we know that we
are at least close on having the full set of ticks.

Literally confirming all of the ticks are present is likely engineering
overkill for very little value.  Everything _should_ be present if the
exit tick is.