Make sure our OHLC data includes the exit tick.
This simplifies matters a bit by doing all of the work in a single
function which can be called from the various American expiries types.

The basic idea is to make sure that we both have the exit tick.. and
that it is included in our OHLC data. There is a small margin for error,
but hopefully it will never come up.