Enable caching by shortcode for simple_contract_info
This code is still bothersome, but it should be faster now in the
majority of cases.  We hold the results for a day to manage constant
reloading. If we happen to have multiple clients with the same contract
and language, they all benefit.  On the other hand, this will likely be
stored at the edge, so we lose some benefit there.

Also includes some test changes to document the full range of returns.