Add rate-limiting on virtual sell rate.
Limiting their buy rate was supposed to help bring this down, but it
didn't do as much as we would like.

I imagine we'll set the sells at approximately the same rates as the
buys.  This should not cause a problem for non-abusive users... and
won't be applied to real money customers, anyway.

The batch sell rate can be set even lower, because excessive use
indicates that they are trying to be "at the limit" (whether balance or
portfolio size) all the time and abusing our good will to maintain this
for them.