variable lookback interval.
We will calculate volatility with any amount of ticks we get from redis
and scale the volatility by duration. If we are pricing a 1 hour
contract, we no longer checks if we have 1 hour worth of ticks in the
past, we will calculate the volatility with the ticks that we have.