zakame/get_count_of_redis_keys (#325)
https://trello.com/c/OopWHh3w/6464-2-zakame-getcountofrediskeys-2

* Add BOM::Test::Helper::Redis

Refactor common code in these PRs into a common helper module:

- https://github.com/regentmarkets/binary-websocket-api/pull/2054
- https://github.com/regentmarkets/bom-rpc/pull/1378

* :lipstick: typo

* :lipstick: Explicitly `return undef`

* Use a simpler match to get the keys

Note that this could have also been written as

    (my $keys) = ...;

Which emphasizes the LHS forcing list context on the RHS a bit more, but
went with the `my ($keys)` form instead in case we add a new scalar to
match from the string, e.g. `my ($keys, $expires)`.

See also https://perlmonks.pairsite.com/?node_id=693666

* :lipstick: cut out unneeded blank lines

* Add short perldoc for BOM::Test::Helper::Redis

I considered interspersing the pod in the module code, but that code is
too short...

* :lipstick: typo

* Make Redis test helper follow Test::More semantics

- it is now called `is_within_threshold`
- has a function prototype similar to `isa_ok` from Test::More
- can now take an optional threshold value, falling back to
  MAX_REDIS_KEYS as default

* Use an even better way to get the keys

Build a hash that and extract the keys from it, thanks @binarytom!