zakame/get_count_of_redis_keys (#2054)
https://trello.com/c/OopWHh3w/6464-2-zakame-getcountofrediskeys-2

* Get a count of Redis keys to compare against a threshold

This test runs at the end of all others, so as to get an idea of how
many keys we have on Redis.  We will compare this with a given
threshold (currently zero, as we need to discover this while we do our
CI builds.)

* Fetch Redis keyspace info from all servers, not just pricing

We were getting a Redis handle from one server only prior, so we're not
getting the right keyspace info for all servers.

* Run Redis key check on all TEST_SUITEs

We get only a reading of 10 keys made in `make structure` test, since
that suite only does basic checks and not actual Redis work.  Hence, add
the redis keys script at the end of the tests for all suites, to get a
better idea...

* Set initial threshold for max Redis keys to 30

Based on CI reports in
https://travis-ci.com/regentmarkets/binary-websocket-api/builds/64256187

* Care about counting Redis keyspace on multiple databases

In the Redis context, a service can have one or more databases, each
with their own keyspace info.  Hence, take care about that and compare
our given threshold against the counts on each database.

* :lipstick: Add more clear comment on Redis INFO keyspace output

So the reader has a better idea about the next line parsing that output.

* :lipstick: `my @redis`, not `our`s

No need for a lexical alias to @main::redis here.

* Bump max key threshold to 40

Per @binarytom at the Backend meeting, thanks!

* Refactor to use BOM::Test::Helper::Redis

Move out common code into a helper module in bom-test.

* :lipstick: Re-add Test::More

Because of `done_testing`

* :lipstick: explicit parens on test_max_redis_keys()

Fell into the indirect object notation trap here.

`no indirect` would have been fine, but costs 2 lines.

* Use `Module->$function()` form to keep `strict 'refs'`

@leonerd++

* Add BOM::Test in structure test

We now use BOM::Test::Helper::Redis here.

* `test_max_redis_keys` is now `is_within_threshold`

Use the refactored helper with closer semantics to Test::More.