zakame/get_count_of_redis_keys (#1378)
https://trello.com/c/OopWHh3w/6464-2-zakame-getcountofrediskeys-2

* Get a count of Redis keys to compare against a threshold

This test runs at the end of all others, so as to get an idea of how
many keys we have on Redis.  We will compare this with a given
threshold (currently zero, as we need to discover this while we do our
CI builds.)

* :lipstick: Remove extra test

* Set initial max keys threshold to 40

Per @binarytom at the backend meeting, thanks!

* Truly run the Redis keys check at the end

It seems that `prove t/BOM` will run the Redis keys test somewhere near
the beginning of the test run, so force that to the end instead.

* Refactor to use BOM::Test::Helper::Redis

Move out common code into a helper module in bom-test.

* :lipstick: Re-add Test::More

Because of `done_testing`...

* :lipstick: explicit parens on test_max_redis_keys()

Fell into the indirect object notation trap here.

`no indirect` would have been fine, but costs 2 lines.

* Use `Module->$function()` form to keep `strict 'refs'`

@leonerd++

* `test_max_redis_keys` is now `is_within_threshold`

Use the refactored helper with closer semantics to Test::More.