Add keywords/triggers for localized status messages
Use `BOM::Platform::Context::localize` on a hash of preset status
messages.  Keywords from this hash can then be passed to the `--message`
option to show the correct translation of the status message.

If the keyword is unrecognized (or in quotes,) just pass it directly on
to the status message setting as a fallback.