Torsten/no temp table (#449)
* We can't call set_staff from BOM::Database::Rose::DB because this
is used for all databases not only clientdb. If we do it in
BOM::Database::ClientDB like it is done now, set_staff will be called
quite often. But since this is used only from BO and the set_staff
implementation itself is just setting a GUC variable, it should not
cause any problems.