Tweak forget_check_redis_sub timeout (#359)
https://trello.com/c/sqmpFf91/6771-5-zakame-tweaktimeoutoncreateproposalawaits-5

* Reduce forget_check_redis_sub timeout to 3 seconds

Revisiting https://trello.com/c/e8av01rA, this pulls down the timeout to
as close to the default 2 seconds in await.pm, per testing in devbox.

* Bump to 5

3 seconds passes on devbox, but not on Travis.

* Capture DataDog timings on forget_check_redis_sub

This reveals some interesting metrics on the devbox: it takes almost 3s
to complete the RPC call on the second `proposal`.  Wondering about the
Travis build timing for this...

* Capture content of statsd tags as well

Further check about the tags in the RPC calls.

* Add a few more pricer subscriptions to make based on contract type

Get even more timing stats by making pairs of CALL/PUT proposals for
each symbol.