fixes app_markup_dollar_amount for non-binary
Currently it is all messed up. App markup is a percentage of payout in
binary contracts and payout is hard-coded to zero in non-binaries. So
basically we are not payout any commission for third party who's buying
non-binaries from us.

We now calculate app markup for non-binary as a function of buy price
instead of payout