Fix MyAffiliates reports generation (#1191)
* Let DocumentUpload accept alternate S3 configuration via import()

This takes advantage of `use` semantics to let DocumentUpload be usable
in other Backoffice scripts that need to upload to S3, e.g.

    use BOM::Platform::Config;
    use BOM::Backoffice::Script::DocumentUpload
        -config => BOM::Platform::Config::third_party->{myaffiliates};

A better approach would be to refactor DocumentUpload entirely as a
shared object; this would be revisited in a future card/sprint.

* Use DocumentUpload to push myaffiliate reports to S3

Instead of attaching the (large) report files to email, upload the
reports to S3 and attach the URL links to the email.

* Remove the (inner) fork for myaffiliate report generator

The double-fork here only hides errors thrown by the child daemon that
does the report generation and email work, which always causes the
parent CGI process to return successfully (despite issues from the
child.)

Note that by removing the inner fork, the whole process is now
synchronous.

* fetch_myaffiliate_payment: supply a filehandle for MD5 checksumming

Use Path::Tiny to get a handle supplying requirements for both
Digest::MD5 and basename.

* DocumentUpload: use previously-set scalars for config

Since the config can now point to an alternate set with different keys
for the S3 credentials, use the previously-set scalars for $access_key,
$secret_key, and $bucket directly to pass to Net::Async::Webservice::S3.

* fetch_myaffiliate_payment: note about short lifetime of reports

Since DocumentUpload enforces a default lifetime of 10 minutes, note
that here for requestors.  We may extend this lifetime on a later
refactor.

* :lipstick: Add spacing for the S3 links in the mail

Make these a bit easier to download... though now considering to aggregate
all files into a so only a single upload/download...

* :lipstick: remove print() of report URL on STDOUT

* Use new PaymentToAccountManager->get_csv_zip helper

Instead of uploading reports to S3 individually, zip them up using the
new method in bom-myaffiliates then upload that one archive.  This
conserves bandwidth and should make it easier for the marketing team to
use this feature.

* fetch_myaffiliate_payment: get rid of forks entirely

There is no longer a need to fork since the underlying bom-myaffiliates
is synchronous, and the S3 upload logic can spin out async on it own
without this script needing to mind it.  Thanks @binarytom for pointing
this out.

* DocumentUpload: allow get_s3_url() with custom expiry

Pass an optional expiry time in seconds along with the required file
name, e.g.

    my $url = get_s3_url($file_name, 3600); # S3 URL with 1h expiry

* fetch_myaffiliate_payment: generate S3 links with 1 hour expiry

Since the reports may be shared to external parties subscribed to the
affiliates@binary.com group, ensure that we have a longer expiry time
for the generated Zip file (at 1 hour) and clarify the message sent in
the mail.