`asset_index` test fixes (#1605)
* 40_pricing.t: fix `is{,_deeply}()` params ordering

`is()` tests are generally of the form `is($got, $expected, $desc)`, but
it was the other way around here, resulting on some temporary confusion.

* 40_pricing.t: simplify result counting

Use a `$count` to capture the number of `@$result`s for the test calls.

* 40_pricing.t: expand the result response test for items

Make this test a little more descriptive on the contents of the
response, as well as accommodate in whatever order the list of available
contract types are given for each asset (via a bag comparison.)

* 40_pricing.t: refactor out common tests into helper sub

Add a `_test_asset_index()` the brings all the common tests together, so
that parameters and expected results can just be plugged into to this.

* Tidy and rename 40_pricing.t to 40_asset_index.t

The test contains only a direct reference to the asset_index RPC
call (and only indirectly as a bom-pricing method) so it seems to make
more sense to rename this test script appropriately.  This cuts down a
few lines as well.

* 40_asset_index.t: let _test_asset_index() return coderef

Realize that _test_asset_index() can just be supplied to subtests
directly, simply by being able to return a coderef (instead of being
wrapped by one.)  This makes it neater and cuts down extra lines.