Mwm/qsort (#567)
* Reduce method calls for redis

* Add simple copy from keys to index

* Add initial queue sorting

* Add attribute for stats tagging

* Add self-reindexing sweep

We don't want to let this go on forever pricing things we might have
lost but not noticed.  On the other hand we don't want to lose all of
the value of having contructed the queue in the first place.

At (presently) about 12h, we strike a balance.  Once things are running
in production we can review whether this makes sense.. and it can be set
from outside, so devops can tune as well.

* Use real_money adjustment when included

* Immprove documentation of build for future

Also helps to explain why things are done as they are today, a tiny
bit.. but most of the intent revolves around making sure future
improvements remain obvious

* Remove priority queue testing

The concept itself is deprecated, no point in confounding the next poor
soul to venture in here.