Include new channels in index (#2968)
* Include new channels in index

* Include real_money value in pricer key

Restricted over the sample code from which is was cribbed to allow more
consloidation of keys

It is nicer in some other places if this key didn't exist, since the
landing compnay info is already there.  However, it was requersted and I
assume someone knows better than I what we'll need.