Import failing code from Pricer Queue Sorting (#516)
https://trello.com/c/ofI69XwQ/9501-mwm-sortpq

* Original code from mwm/qsort (#509)

I'm really unhappy with this proliferation of checks and values.
However, I cannot make the changes I want to see today.  I'll spend some
capital on that later.

* Improve semantics for finding real_money

I mean, it's what people asked me to do.. but I am not sure they
understood it was this mocked-up struct.

* Expand allowed key threshold to 50

We have an extra key in the pricer redis for the channel index now
which was tripping on this bound.

Since I don't think the actual count is a required hard limit, giving a
little extra breathing room for the next personn to add a key or two