Pricer Queue Sorting (#584)
https://trello.com/c/ofI69XwQ/9501-mwm-sortpq

* Original code from Mwm/qsort (#567)

* Reduce method calls for redis

* Add simple copy from keys to index

* Add initial queue sorting

* Add attribute for stats tagging

* Add self-reindexing sweep

We don't want to let this go on forever pricing things we might have
lost but not noticed.  On the other hand we don't want to lose all of
the value of having contructed the queue in the first place.

At (presently) about 12h, we strike a balance.  Once things are running
in production we can review whether this makes sense.. and it can be set
from outside, so devops can tune as well.

* Use real_money adjustment when included

* Immprove documentation of build for future

Also helps to explain why things are done as they are today, a tiny
bit.. but most of the intent revolves around making sure future
improvements remain obvious

* Remove priority queue testing

The concept itself is deprecated, no point in confounding the next poor
soul to venture in here.

* Remove deprecated high priority queue

* Ignore vim swap files

* Improve visibility for testing

* Use better indicator of extant contracts

* Expand and improve testing around queueing

* Improve variable naming

It was far too difficult to hold all that state in one's head when the
names can do some of the heavy lifting.

* Improve testability (and testing!) of daemon

* Improve code and testing around scoring

Based on some very helpful code-review comments from Mike Mueller

* Document injectable dependencies for PriceDaemon

* Extract extract_from_channel_key

* Improve log message string formatting

* Refactor `passes_until_reindex'

I'm not sure the difference is worth concerning ourselves, but here it
is anyway.

* Stub out process_job for testing

This exercises a bit more of the code, since we give a vaguely correct
reponse

* FIx various spelling and verbiage issues

Wherein Tom makes up for my horrible keyboarding skills

Co-Authored-By: Tom Molesworth <tom@binary.com>

* Return explicit undef

* Quotemeta on prefix match

* Convert jobs and channels keys to attributes

This might help with testing even if they seem unlikely to change in
production mostly ever.