Import failing code from Pricer Queue Sorting (#3034)
And the hits keep on coming...

* Original code from mwm/qsort (#2968)

* Include new channels in index

* Include real_money value in pricer key

Restricted over the sample code from which is was cribbed to allow more
consloidation of keys

It is nicer in some other places if this key didn't exist, since the
landing compnay info is already there.  However, it was requersted and I
assume someone knows better than I what we'll need.

* Remove buggy 'optimisation'

This was intended to make it such that we could reduce the work done by
the pricers by skipping a pricing cycle when we know that there will be
noone listening. On the other hand, the channel key was *not*
cleaned up here. No pricer daemon would even see it to clean up, either.

We could remove both here, but I am not an ardent fan of method
modifiers.  Also, historically an extra pricing cycle has not been an
issue, so we'll continue with that.  It demonstrably worked.  This
demonstrably did not.