Improve testing realism
Testing pricing_interval

Less stubbing, more "process" interaction.

If numbers herein look magic, they kinda are.  They are just tuned to
process in a reasonable way, but there's no reason not to change them if
you like.

This isn't helping much to find the source of the problems, but it's
still likely useful.