use JSON::MaybeXS (#261)
* use JSON::MaybeXS

* 23

* trigger circleci

* trigger circleci

* fix error

* test again 2020-09-22 06:31:20

* test again 2020-09-28 08:42:21

* change decode_json_text to JSON->new->allow_nonref->decode

The problem is that JSON::MaybeUTF8::decode_json_text will call JSON->decode at
last, but in Cpanel::JSON::XS, the default option 'allow_nonref' is false, and
it is true in JSON. In such case, it will be ok to decode a string with a number
"1234" in JSON, but not in Cpanel::JSON::XS, and then JSON::MaybeUTF8 will also fail.

And in this script some values will be this format "1234". So we must set
allow_non_ref ourselves.