trade_warning/user_limit_email (#1393)
* Stop sending user limit surpass notification emails

* Ignore all user limit crossings. We do not react on them

* comments

* Ignore clients limit crossage bases on fixed limit

New clients limit is set to 2000. Ignore rest of them or anything lower
than this

* Use constants

* Cleanup regex

Remove parans, we don't need capture groups

* Get new clients limit from the db

* Retrigger tests

* Update lib/BOM/Platform/Script/TradeWarnings.pm

Co-authored-by: Greg Sabino Mullane <greg@binary.com>

* Update lib/BOM/Platform/Script/TradeWarnings.pm

Co-authored-by: Greg Sabino Mullane <greg@binary.com>

* Make tidy

* Set $| to 1; Some logs seems to be missing in datadog

We can not see all the logs we expectt in datadog which we suspect
might be duo to buffering.

* Add comments

* cleanup

* Make tidy

* Add logging to trade warning

It's mostly used to make QA process easier.
Also we want to be sure only one mechanism is used to print logs.

* Remove unsed Redis
Change logs

* Add tests

* Make tidy

* Use formatted logging

Co-authored-by: Greg Sabino Mullane <greg@binary.com>