queue proposals for pricing on tick updates (#1443)
* queue proposals for pricing on tick updates

instead of retrieving the list of all the contracts from redis in the
beginning of every second and queueing them all for pricing, we now keep
the list of contracts for pricing in memory grouped by symbol and we
regularly update this list from redis. We also listen for tick updates,
and once the tick has been updated we queue for pricing only contracts
for the corresponding symbol.

* use Finance::Underlying to get the symbol

* add delay after tick before triggering pricing

we have a problem with getting prices for duplicate spots, the reason it
seems that we're getting the notice of tick update and trigger repricing
before the new tick is available in replicated redis for the pricer to
read

* trigger tests [ci] 2023-05-02 19:58:37

* trigger test [ci]

* force exit if there are no ticks for 10 sec

this is to work around the bug in Net::Async::Redis that doesn't handle
disconnects from redis server correctly

* trigger test [ci]