pricing/send-tick-to-pricer (#1456)
* queue proposals for pricing on tick updates

instead of retrieving the list of all the contracts from redis in the
beginning of every second and queueing them all for pricing, we now keep
the list of contracts for pricing in memory grouped by symbol and we
regularly update this list from redis. We also listen for tick updates,
and once the tick has been updated we queue for pricing only contracts
for the corresponding symbol.

* use Finance::Underlying to get the symbol

* add delay after tick before triggering pricing

we have a problem with getting prices for duplicate spots, the reason it
seems that we're getting the notice of tick update and trigger repricing
before the new tick is available in replicated redis for the pricer to
read

* force exit if there are no ticks for 10 sec

this is to work around the bug in Net::Async::Redis that doesn't handle
disconnects from redis server correctly

* remove skipping pricing on duplicate spot

we now only trigger pricing when new tick is available or every 5 sec,
so we always want to price what is in the queue

* remove unused attribute

* submit jobs if there was no ticks for 5 sec

if there was no ticks for symbol for 5 seconds submit all proposals and
contracts for that symbol for pricing.

* tidy futures

* fix reporting of overflow

that affects stats in pricer_daemon_queue_stats redis key,
datadog reporting is correct

* detect if ticks_stayed_in missing the current tick

if we see that ticks_stayed_in doesn't include the current tick we check
if the current ticks broke the barriers and update ticks_stayed_in
accordingly, i.e. either increase the latest ticks_stayed_in or put 0 in
there.

* forward tick to pricer

when triggering pricing forward the tick to pricer_daemon to ensure that
it uses that exact tick for pricing. Otherwise it is possible that it'll
be using the previous tick.

* remove unused code

* track queue latency

* fix detection of disconnects

* remove unnecessary logging statement

* trigger tests [ci] 2023-06-07 08:57:13