make sure the latest tick is included (#2521)
* make sure the latest tick is included

when we're building tick_stream or ticks_for_tick_expiry lists for
accumulator we want to make sure that the current spot is included in
the list. If we pass current spot when we initialize the contract and
that current spot comes from redis pub/sub, it is possible that the spot
hasn't been replicated yet to redis that tick accessor is using to get
the list.

* trigger tests [ci] 2023-06-07 08:57:13