pricing/metrics-review (#1464)
* use a single rpop with count

now that we've upgraded redis in debian_ci image we can do that.

* measure latency of starting pricing

this is time between receiving a tick in price-queue and starting
processing the price for a contract in price-daemon

* measure feed latency and processing time

measure latency between beginning of the second and the time when we get
the tick in price-queue. Also measure averate time required to process
the tick.

* trigger tests [ci] 2023-06-27 10:37:21