update/json-validator (#1186)
* we don't limit the size of passthrough any more

* wsapi/remove-maxsize-param

* fix for new JSON::Validator

it uses different way to pass args to coerce

* update/json-validator

* add a test to check that the correct value is accepted

* debug

* remove test for +100 coercion

+100 is not a valid JSON number format, we can't require JSON::Validator
to coerce it. The primary purpose of JSON::Validator is to validate that
JSON conforms the schema, coercion is mainly to ensure that when we
generate JSON from Perl structures we produce correct types, not to
perform transformation of input data.