pricer/fix-localization (#1524)
* ensure that we always start with the same global state

there is a lot of places where we still call localize while calculating
price. We should stop doing that, but that's a separate task.

* create a copy of the original request

this is a shallow copy, but here we mainly want to preserve longcode.
With adjusted_response being a reference to the original one we were
rewriting longcode with the localized one.