restart workers periodically (#1572)
pricer is slowly leaking memory. The prime suspect is localization
routine, when it encounters unknown strings it adds them into lexicon
hash, if we don't know how to translate a particular longcode, for
example, it may add it with expanded parameters into the lexicon.
Attempt to fix BOM::Platform::Context::I18N was not successfull as the
fix was interfering with implementation in other services, particularly
binary-websocket-api. Hence we decided to simply proceed with periodic
restarts, especially as the source of the leak hasn't been confirmed.