Improve risk profile caching (#3428)
* Ignore ctags

* Improve caching utility

It's unfortunate that there is some unused work to be done in the
per-client custom limits, but its made up in the global limits.

This affects the speed of pricing by reducing the number of
Date::Utility object to create.

Sometimes I really miss "everything is an expression"... including `if`

* Add tests to prove new implementation

This demonstrates the cache adjustments working properly under time
changing conditions.